HSEARCH-3222 Move *SearchProjectionFactoryImpl to the projection package
